requirements.txt
pycairo
webcolors
requests
